requirements.txt
django
djangorestframework
pillow
channels
daphne
channels-redis
whiteNoise
python-dotenv
django-extensions
numpy==1.24.1
django-cors-headers
uvicorn
gunicorn
psycopg2-binary
django-storages[azure]
azure-storage-blob
azure-identity
django-jet-reboot
djangorestframework_simplejwt
qdrant-client
tqdm